tickets/requirements.txt
pika
python-dotenv
rsa
pydantic